@@ -11,36 +11,49 @@ namespace App\Handler;
 use App\Entity\Address;
 use App\Entity\Elector;
 use App\Repository\ElectorRepository;
+use App\Service\CsvReader;
 use Doctrine\ORM\EntityManagerInterface;
-use League\Csv\Reader;
 
-class ElecListCsvRecordHandler
+class ElecListCsvImporter
 {
-    /**
-     * @var EntityManagerInterface
-     */
-    private $entityManager;
-    /**
-     * @var ElectorRepository
-     */
-    private $electorRepository;
+    /** @var EntityManagerInterface */
+    private EntityManagerInterface $entityManager;
+
+    /** @var ElectorRepository */
+    private ElectorRepository $electorRepository;
+
+    /** @var CsvReader */
+    private CsvReader $csvReader;
+
+    /** @var array  */
+    private array $params;
 
     /**
      * @param EntityManagerInterface $entityManager
      * @param ElectorRepository $electorRepository
+     * @param CsvReader $csvReader
+     * @param array $params
      */
-    public function __construct(EntityManagerInterface $entityManager, ElectorRepository $electorRepository)
-    {
+    public function __construct(
+        EntityManagerInterface $entityManager,
+        ElectorRepository $electorRepository,
+        CsvReader $csvReader,
+        array $params
+    ) {
         $this->entityManager = $entityManager;
         $this->electorRepository = $electorRepository;
+        $this->csvReader = $csvReader;
+        $this->params = $params;
     }
 
-    public function importFile(Reader $file)
+    public function importFile(String $filePath): bool
     {
-        $count = $file->count();
+        $records = $this->csvReader->getRecords($filePath);
+        $count = $this->csvReader->getReader()->count();
+
         $counter = 0;
 
-        foreach ($file->getRecords() as $key => $record) {
+        foreach ($records as $key => $record) {
             $this->saveRecord($record);
 
             if ($counter === 1000 || $key === $count) {
@@ -51,6 +64,8 @@ class ElecListCsvRecordHandler
 
             $counter += 1;
         };
+
+        return true;
     }
 
     public function clear()
@@ -75,10 +90,10 @@ class ElecListCsvRecordHandler
 
         $elector = new Elector();
         $elector
-            ->setFirstname(trim($record['prénoms']))
+            ->setFirstname(trim($record['firstname']))
             ->setLastname(trim($lastName))
-            ->setBirthname(trim($record['nom de naissance']))
-            ->setVoteOffice(trim($record['code du bureau de vote']));
+            ->setBirthname(trim($record['birthname']))
+            ->setVoteOffice(trim($record['vote_office']));
 
         return $elector;
     }
@@ -87,22 +102,22 @@ class ElecListCsvRecordHandler
     {
         $address = new Address();
         $address
-            ->setAdd1(trim($record['complément 1']))
-            ->setAdd2(trim($record['complément 2']))
-            ->setStreet(trim($record['libellé de voie']))
-            ->setCity(trim($record['commune']))
-            ->setPostcode(trim($record['code postal']))
-            ->setNumber(trim($record['numéro de voie']));
+            ->setAdd1(trim($record['add1']))
+            ->setAdd2(trim($record['add2']))
+            ->setStreet(trim($record['result_name']?? $record['street']))
+            ->setCity(trim($record['result_city'] ?? $record['city']))
+            ->setPostcode(trim($record['result_postcode'] ?? $record['postcode']))
+            ->setNumber(trim($record['result_housenumber'] ?? $record['house_number']));
 
         return $address;
     }
 
     private function getLastName(array $record)
     {
-        if ($record['nom d\'usage']) {
-            return $record['nom d\'usage'];
+        if ($record['lastname']) {
+            return $record['lastname'];
         }
 
-        return $record['nom de naissance'];
+        return $record['birthname'];
     }
 }
